+++ test/td_bg/business_concepts/business_concepts_test.exs
@@ -12,7 +12,7 @@ defmodule TdBg.BusinessConceptsTest do
   alias TdCache.I18nCache
   alias TdCache.Redix
   alias TdCache.Redix.Stream
-  alias TdCore.Search.IndexWorkerMock
+  alias TdCore.Search.IndexWorker
   alias TdDfLib.Format
   alias TdDfLib.RichText
 
@@ -51,7 +51,7 @@ defmodule TdBg.BusinessConceptsTest do
   setup context do
     on_exit(fn ->
       Redix.del!(@stream)
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
     end)
 
     case context[:template] do
@@ -697,7 +697,7 @@ defmodule TdBg.BusinessConceptsTest do
   describe "update_business_concept_version/2" do
     @tag template: @content
     test "updates the business_concept_version if data is valid and publishes an event to the audit stream" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       %{user_id: user_id} = build(:claims)
 
       %{id: parent_id, parent_id: root_id} = parent = insert(:domain, parent: build(:domain))
@@ -752,12 +752,12 @@ defmodule TdBg.BusinessConceptsTest do
                Jason.decode!(payload)
 
       assert_lists_equal(domain_ids, [root_id, parent_id, domain_id])
-      assert [{:reindex, :concepts, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "updates the content with valid content data" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       %{id: domain_id} = insert(:domain)
 
       template_name = "test_template"
@@ -831,8 +831,8 @@ defmodule TdBg.BusinessConceptsTest do
       assert %BusinessConceptVersion{} = new_business_concept_version
       assert new_business_concept_version.content["Field1"] == "New first field"
       assert new_business_concept_version.content["Field2"] == "Second field"
-      assert [{:reindex, :concepts, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "updates the content with valid i18n_content data" do
@@ -1094,7 +1094,7 @@ defmodule TdBg.BusinessConceptsTest do
 
     @tag template: @content
     test "get_business_concept_version/2 returns the business_concept_version by concept id and version" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       claims = build(:claims)
       %{id: domain_id} = insert(:domain)
 
@@ -1132,8 +1132,8 @@ defmodule TdBg.BusinessConceptsTest do
 
       refute BusinessConcepts.get_business_concept_version(business_concept_id + 1, id)
 
-      assert [_, _, _] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [_, _, _] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "get_business_concept_version/2 returns preloads" do