+++ test/td_bg/business_concepts/workflow_test.exs
@@ -8,7 +8,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
   alias TdBg.BusinessConcepts.Workflow
   alias TdCache.Redix
   alias TdCache.Redix.Stream
-  alias TdCore.Search.IndexWorkerMock
+  alias TdCore.Search.IndexWorker
 
   @stream TdCache.Audit.stream()
   @template_name "test_template"
@@ -22,7 +22,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
   setup do
     on_exit(fn ->
       Redix.del!(@stream)
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
     end)
 
     [claims: build(:claims)]
@@ -74,7 +74,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
     end
 
     test "creates a new published version and the previous version remains the current" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       business_concept_version =
         insert(:business_concept_version, status: "published", current: true, type: @template_name)
@@ -83,8 +83,8 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
 
       assert %{current: %{current: false, business_concept_id: id}} = res
       assert BusinessConcepts.get_business_concept_version!(business_concept_version.id).current
-      assert IndexWorkerMock.calls() == [{:reindex, :concepts, [id]}]
-      IndexWorkerMock.clear()
+      assert IndexWorker.calls() == [{:reindex, :concepts, [id]}]
+      IndexWorker.clear()
     end
 
     test "creates a new version and copies the identifier from the previous version one", %{
@@ -92,7 +92,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       template_with_identifier: template_with_identifier,
       identifier_name: identifier_name
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       existing_identifier = "00000000-0000-0000-0000-000000000000"
       concept = build(:business_concept, %{type: template_with_identifier.name})
 
@@ -112,12 +112,12 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
                }
              } = res
 
-      assert IndexWorkerMock.calls() == [{:reindex, :concepts, [id]}]
-      IndexWorkerMock.clear()
+      assert IndexWorker.calls() == [{:reindex, :concepts, [id]}]
+      IndexWorker.clear()
     end
 
     test "creates a new published version and advance to a publish state will make it current" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       business_concept_version =
         insert(:business_concept_version, status: "published", current: true, type: @template_name)
@@ -135,12 +135,12 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       refute BusinessConcepts.get_business_concept_version!(business_concept_version.id).current
       assert BusinessConcepts.get_business_concept_version!(res.published.id).current
 
-      assert [_, _, _] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [_, _, _] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event to the audit stream" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       business_concept_version =
         insert(:business_concept_version, status: "published", type: @template_name)
@@ -149,15 +149,15 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
                Workflow.new_version(business_concept_version, %Claims{user_id: 1234})
 
       assert {:ok, [%{id: ^event_id}]} = Stream.read(:redix, @stream, transform: true)
-      assert [_] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [_] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event to the audit stream with domain ids", %{
       concept: concept,
       domain_ids: domain_ids
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       business_concept_version =
         insert(:business_concept_version,
@@ -172,8 +172,8 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
                Stream.read(:redix, @stream, transform: true)
 
       assert %{"domain_ids" => ^domain_ids} = Jason.decode!(payload)
-      assert [{:reindex, :concepts, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end
 
@@ -181,7 +181,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
     setup :create_concept_with_parents
 
     test "changes the status and audit fields" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       %{last_change_at: ts} =
         business_concept_version = insert(:business_concept_version, status: "draft")
@@ -194,8 +194,8 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
                published
 
       assert DateTime.diff(last_change_at, ts, :microsecond) > 0
-      assert [{:reindex, :concepts, _}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, _}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{
@@ -203,7 +203,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       concept: concept,
       domain_ids: domain_ids
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       business_concept_version =
         insert(:business_concept_version, status: "draft", business_concept: concept)
@@ -213,20 +213,20 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       assert %{id: ^event_id, payload: payload} = event
       assert %{"domain_ids" => ^domain_ids} = Jason.decode!(payload)
 
-      assert [{:reindex, :concepts, _}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, _}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event including subscribable_fields", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       business_concept_version = insert(:business_concept_version, status: "draft")
 
       assert {:ok, %{audit: event_id}} = Workflow.publish(business_concept_version, claims)
       assert {:ok, [event]} = Stream.read(:redix, @stream, transform: true)
       assert %{id: ^event_id, payload: payload} = event
       assert %{"subscribable_fields" => _} = Jason.decode!(payload)
-      assert [{:reindex, :concepts, _}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, _}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end
 
@@ -234,7 +234,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
     setup :create_concept_with_parents
 
     test "rejects business_concept", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       reason = "Because I want to"
       business_concept_version = insert(:business_concept_version, status: "pending_approval")
 
@@ -244,8 +244,8 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       assert %{status: "rejected", reject_reason: ^reason, business_concept_id: concept_id} =
                business_concept_version
 
-      assert [{:reindex, :concepts, [concept_id]}] == IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, [concept_id]}] == IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{
@@ -253,7 +253,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       concept: concept,
       domain_ids: domain_ids
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       reason = "Because I want to"
 
       business_concept_version =
@@ -264,8 +264,8 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       assert {:ok, [event]} = Stream.read(:redix, @stream, transform: true)
       assert %{id: ^event_id, payload: payload} = event
       assert %{"domain_ids" => ^domain_ids} = Jason.decode!(payload)
-      assert [{:reindex, :concepts, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end
 
@@ -273,17 +273,17 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
     setup :create_concept_with_parents
 
     test "updates the business_concept", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       %{user_id: user_id} = claims
       business_concept_version = insert(:business_concept_version)
 
       assert {:ok, %{updated: %{business_concept_id: id} = business_concept_version}} =
                Workflow.submit_business_concept_version(business_concept_version, claims)
 
-      assert IndexWorkerMock.calls() == [{:reindex, :concepts, [id]}]
+      assert IndexWorker.calls() == [{:reindex, :concepts, [id]}]
 
       assert %{status: "pending_approval", last_change_by: ^user_id} = business_concept_version
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{
@@ -291,7 +291,7 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       concept: concept,
       domain_ids: domain_ids
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       business_concept_version = insert(:business_concept_version, business_concept: concept)
 
       assert {:ok, %{audit: event_id}} =
@@ -300,8 +300,8 @@ defmodule TdBg.BusinessConcepts.WorkflowTest do
       assert {:ok, [event]} = Stream.read(:redix, @stream, transform: true)
       assert %{id: ^event_id, payload: payload} = event
       assert %{"domain_ids" => ^domain_ids} = Jason.decode!(payload)
-      assert [{:reindex, :concepts, _}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :concepts, _}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end